+++ Cult of the Lamb/TraitControl/Configs.cs
@@ -5,12 +5,18 @@ public partial class Plugin
     // Trait Replacement settings
     internal static ConfigEntry<bool> NoNegativeTraits { get; private set; }
     internal static ConfigEntry<bool> UseUnlockedTraitsOnly { get; private set; }
+    internal static ConfigEntry<bool> PreferExclusiveCounterparts { get; private set; }
     internal static ConfigEntry<bool> IncludeImmortal { get; private set; }
     internal static ConfigEntry<bool> IncludeDisciple { get; private set; }
+    internal static ConfigEntry<bool> IncludeDontStarve { get; private set; }
+    internal static ConfigEntry<bool> IncludeBlind { get; private set; }
+    internal static ConfigEntry<bool> IncludeBornToTheRot { get; private set; }
     internal static ConfigEntry<bool> ShowNotificationsWhenRemovingTraits { get; private set; }
     internal static ConfigEntry<bool> ShowNotificationsWhenAddingTraits { get; private set; }
 
     // Trait Weights settings
     internal static ConfigEntry<bool> EnableTraitWeights { get; private set; }
+    internal static ConfigEntry<bool> UseAllTraits { get; private set; }
     internal static Dictionary<FollowerTrait.TraitType, ConfigEntry<float>> TraitWeights { get; } = new();
+    internal static List<FollowerTrait.TraitType> AllTraitsList { get; } = [];
 }