+++ Cult of the Lamb/TraitControl/GlobalUsings.cs
@@ -6,6 +6,7 @@ global using System.Collections.Generic;
 global using System.IO;
 global using System.Linq;
 global using System.Reflection;
+global using System.Reflection.Emit;
 global using BepInEx;
 global using BepInEx.Configuration;
 global using BepInEx.Logging;